requirements.txt
Python:
nanoatp 
python-telegram-bot
arrow

Java:
https://github.com/unbescape/unbescape/
